+++ plume/src/main/kotlin/io/github/plume/oss/drivers/TigerGraphDriver.kt
@@ -28,9 +28,10 @@ import io.github.plume.oss.util.ExtractorConst.BOOLEAN_TYPES
 import io.github.plume.oss.util.ExtractorConst.INT_TYPES
 import io.github.plume.oss.util.PlumeKeyProvider
 import io.shiftleft.codepropertygraph.generated.EdgeTypes
+import io.shiftleft.codepropertygraph.generated.NodeTypes.META_DATA
+import io.shiftleft.codepropertygraph.generated.NodeTypes.UNKNOWN
 import io.shiftleft.codepropertygraph.generated.PropertyNames
 import io.shiftleft.codepropertygraph.generated.PropertyNames.*
-import io.shiftleft.codepropertygraph.generated.NodeTypes.*
 import io.shiftleft.codepropertygraph.generated.nodes.NewMetaDataBuilder
 import io.shiftleft.codepropertygraph.generated.nodes.NewNodeBuilder
 import khttp.responses.Response
@@ -258,13 +259,9 @@ class TigerGraphDriver internal constructor() : IOverridenIdDriver, ISchemaSafeD
         val vDels = mutableListOf<DeltaGraph.VertexDelete>()
         val eDels = mutableListOf<DeltaGraph.EdgeDelete>()
         PlumeTimer.measure(ExtractorTimeKey.DATABASE_READ) {
-            dg.changes.asSequence().filterIsInstance<DeltaGraph.VertexAdd>().map { it.n }
+            dg.changes.filterIsInstance<DeltaGraph.VertexAdd>().map { it.n }
                 .filterNot(::exists)
-                .map {
-                    if (it.id() < 0) it.id(PlumeKeyProvider.getNewId(this))
-                    it
-                }.distinctBy { it.id() }
-                .toCollection(vAdds)
+                .forEachIndexed { i, va -> if (vAdds.none { va === it }) vAdds.add(va.id(-(i + 1).toLong())) }
             dg.changes.filterIsInstance<DeltaGraph.EdgeAdd>().distinct()
                 .filterNot { exists(it.src, it.dst, it.e) }
                 .toCollection(eAdds)
@@ -275,8 +272,19 @@ class TigerGraphDriver internal constructor() : IOverridenIdDriver, ISchemaSafeD
         }
         PlumeTimer.measure(ExtractorTimeKey.DATABASE_WRITE) {
             // Aggregate all requests going into the add
-            val vAddPayload = vAdds.map(::createVertexPayload)
-                .foldRight(mutableMapOf<Any, Any>()) { x, y -> deepMerge(x as MutableMap<Any, Any>, y) }
+            val vAddPayload = mapOf("CPG_VERT" to vAdds
+                .map {
+                    Pair(
+                        it.id(PlumeKeyProvider.getNewId(this)).id().toString(),
+                        CollectionConverters.MapHasAsJava(it.build().properties()).asJava() +
+                                mapOf("label" to it.build().label())
+                    )
+                }
+                .foldRight(mutableMapOf<String, Any>()) { x, y ->
+                    y.apply {
+                        this[x.first] = extractAttributesFromMap(x.second.toMutableMap())
+                    }
+                })
             val eAddPayload = eAdds.map { createEdgePayload(it.src, it.dst, it.e) }
                 .foldRight(mutableMapOf<Any, Any>()) { x, y -> deepMerge(x as MutableMap<Any, Any>, y) }
             if (vAdds.size > 1 || eAdds.size > 1) {
@@ -485,7 +493,7 @@ class TigerGraphDriver internal constructor() : IOverridenIdDriver, ISchemaSafeD
                 val vertices = o["allVert"] as JSONArray
                 vertices.map { vertexPayloadToNode(it as JSONObject) }.forEach {
                     val n = it.build()
-                    if (graph.node(it.id()) == null) {
+                    if (!vs.containsKey(it.id())) {
                         val node = graph.addNode(it.id(), n.label())
                         n.properties().foreachEntry { key, value -> node.setProperty(key, value) }
                         vs[it.id()] = node
@@ -870,7 +878,7 @@ CREATE QUERY deleteMethod(STRING FULL_NAME) FOR GRAPH <GRAPH_NAME> SYNTAX v2 {
   allVert = start;
   # Get method's body vertices
   start = SELECT t
-          FROM start:s -((_AST>|_REF>|_CFG>|_ARGUMENT>|_CAPTURED_BY>|_BINDS_TO>|_RECEIVER>|_CONDITION>|_BINDS>)*) - :t;
+          FROM start:s -((_AST>|_CONTAINS>|_REF>|_CFG>|_ARGUMENT>|_CAPTURED_BY>|_BINDS_TO>|_RECEIVER>|_CONDITION>|_BINDS>)*) - :t;
   allVert = allVert UNION start;
 
   DELETE s FROM allVert:s;